refactor: rename module
Rename the module for running child processes to "child-process" to better reflect this.
