+++ src/io/upm-config-io.ts
@@ -7,7 +7,7 @@ import { ReadTextFile, WriteTextFile } from "./fs-result";
 import { tryGetEnv } from "../utils/env-util";
 import { StringFormatError, tryParseToml } from "../utils/string-parsing";
 import { tryGetWslPath, WslPathError } from "./wsl";
-import { ChildProcessError, RunChildProcess } from "./process";
+import { ChildProcessError, RunChildProcess } from "./child-process";
 import { GetHomePath } from "./special-paths";
 import { GenericIOError } from "./common-errors";
